Improve safety of load_build_file()
Before, load_build_file() had to be used in a try/except, due to not handling its own exceptions.
Now, load_build_file() will raise exceptions if build file isn't a Python file, as well as if
the user specifies a build file that's not named 'build.py' and it can't be found.
In both cases, Pixis should stop.

If we're looking for 'build.py' and it's not there, it's possible that the user wants to use
Pixis defaults, so we can continue in that case.